Fixed compiler error, and cleared overflow warning for  src/ByteArray.h

+++ src/ByteArray.h
@@ -24,7 +24,7 @@
 namespace Ollie {
     namespace OllieBuffer {
 
-        static const int nPos = std::string::npos;
+        static const size_t nPos = std::string::npos;
 
         class ByteArray {
